service/web/requirements.txt
Flask
gunicorn
psycopg2-binary
py-dateutil
pytest
pytz
requests
